backend/requirements.txt
Flask
Flask-Cors
Pillow
pytesseract
opencv-python
